Drop -Wswitch-default and mask flag arrays as unsigned
-Wswitch-default reported 657 switches for lacking an empty default
label. -Wswitch in the baseline already reports unhandled enum values
and -Wswitch-enum stays in the migration tier, so the flag added no
correctness signal.

IS_SET_AR masked an int array element with the unsigned Q_BIT value,
converting the element at every use of IS_NPC, GET_NAME, AFF_FLAGGED,
PRF_FLAGGED and the colour macros. Cast the element explicitly; the
expression type and value are unchanged.

Clang 22 sign-conversion falls from 15028 to 3385 sites.

+++ src/utils.h
@@ -796,7 +796,7 @@ void char_from_furniture(struct char_data *ch);
 #define Q_BIT(x) (1U << ((x) % 32))
 
 /** 1 if bit is set in the bitarray represented by var, 0 if not. */
-#define IS_SET_AR(var, bit) ((var)[Q_FIELD(bit)] & Q_BIT(bit))
+#define IS_SET_AR(var, bit) ((unsigned int)(var)[Q_FIELD(bit)] & Q_BIT(bit))
 
 /** Set a specific bit in the bitarray represented by var to 1. */
 #define SET_BIT_AR(var, bit) ((var)[Q_FIELD(bit)] |= Q_BIT(bit))
